Pure_JavaScript_News_APP/requirements.txt
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
newsapi-python==0.2.6
requests==2.22.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
